Do not try to back up extended attributes
While it would be handy to be able to back up extended attributes, there
are too many problems that would result in trying to do so on an
rdiff-backup server running as unprivileged user -- especially under
SELinux enforcement. When performing a restore, the correct SELinux
attributes would be applied anyway during the write operation, or can be
fixed post-restore by running restorecon.

Passing --no-eas tells rdiff-backup not to try setting extended
attributes on files created on the remote backup server.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>